+++ clevertap-core/src/test/java/com/clevertap/android/sdk/cryption/CryptHandlerTest.kt
@@ -2,18 +2,20 @@ package com.clevertap.android.sdk.cryption
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.clevertap.android.sdk.Constants
-import com.clevertap.android.sdk.Constants.AES_PREFIX
-import com.clevertap.android.sdk.Constants.AES_SUFFIX
+import com.clevertap.android.sdk.cryption.CryptHandler.EncryptionAlgorithm
 import io.mockk.MockKAnnotations
 import io.mockk.confirmVerified
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
+import io.mockk.mockkObject
+import io.mockk.unmockkObject
 import io.mockk.verify
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Before
+import org.junit.Ignore
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -61,6 +63,7 @@ class CryptHandlerTest {
     }
 
     @Test
+    @Ignore("We have made sure to never call this method on AES legacy encrypted text, so this use case is not needed")
     fun `encryptSafe - already AES encrypted text - returns same text without re-encryption`() {
         val encryptedText = "${Constants.AES_PREFIX}alreadyEncrypted${Constants.AES_SUFFIX}"
 
@@ -235,6 +238,34 @@ class CryptHandlerTest {
         val invalidAesGcmEncryptedText = "${Constants.AES_GCM_PREFIX}testEncryptedText"
 
         assertFalse(CryptHandler.isTextAESGCMEncrypted(invalidAesGcmEncryptedText))
+    }
+
+    @Test
+    fun encryptSafe_defaultValues_checkCorrectEncryptionTypes() {
+        val aesGcmEncryptedText = "${Constants.AES_GCM_PREFIX}testEncryptedText${Constants.AES_GCM_SUFFIX}"
+
+        mockkObject(CryptHandler.Companion)
+
+        cryptHandler.encryptSafe(aesGcmEncryptedText)
+        verify { CryptHandler.isTextAESGCMEncrypted(aesGcmEncryptedText) }
+        verify(exactly = 0) { cryptFactory.getCryptInstance(EncryptionAlgorithm.AES_GCM) }
+        verify(exactly = 0) { crypt.encryptInternal(aesGcmEncryptedText) }
+
+        unmockkObject(CryptHandler.Companion)
+    }
+
+    @Test
+    fun decryptSafe_defaultValues_checkCorrectEncryptionTypes() {
+        val testPlainText = "testPlainText"
+
+        mockkObject(CryptHandler.Companion)
+
+        cryptHandler.decryptSafe(testPlainText)
+        verify { CryptHandler.isTextAESGCMEncrypted(testPlainText) }
+        verify(exactly = 0) { cryptFactory.getCryptInstance(EncryptionAlgorithm.AES_GCM) }
+        verify(exactly = 0) { crypt.decryptInternal(testPlainText) }
+
+        unmockkObject(CryptHandler.Companion)
     }
 
 }